Updated CharacterController and world
The CharacterController now is able to be created on a slot with an "attachcharactercontroller" method. This will be replaced with KodEngine components later.
Session switching is now possible with tab, with each character controller being isolated to their own session. It is a known bug that disabled character controllers will still listen for jump inputs