Created a Custom component system
Components no longer inherit Unity MonoBehavior. Instead, they inherit a Component class containing a few methods. Components can now use actions to listen to the focus state of the world they are present in. The CharacterController component will not process inputs or physics if not focused in the world the component is present it.