Added support for a simple material
PBS_Metallic now properly works, with an albedo color and texture.
Assigning a URI to a Texture2D component will load the texture properly.
Assigning a Texture2D to a PBS_Metallic texture property will have it appear ingame.
If a MeshRenderer has no material, the BuiltInMaterial is used instead.