Added MeshCollider and fixed floor creation
The MeshCollider component can take in a mesh and will create a collider on the gameobject it was attached to. The mesh can be changed during runtime, and the collider will update

The demo scene originially called a function to ceate a floor for the player to stand on. This floor used Unity components as mesh colliders didn't exist prior. This function now uses KodEngine components.

When building the scene, 100% of the functionality in the viewport is done with attached KodEngine components, meaning that if a KodEngine scene inspector existed, the scene could be created ingame. That's pretty cool