Added support for building
Previously, the shaders would load from the unityeditor. This would allow it to work in the preview, but building would fail as it wasn't packed with those shaders. Every unity 2021.3.5f1 editor shader has been added to the resources folder. The project now builds in its current state and behaves as expected.