Further networking work
I've successfully gotten network messages to work, as well as serializing objects, however I have ran into a bit of a snag. Deserializing the objects when they themselves reference the same type will not work. I plan on instead having RefID objects attached where the objects themselves would be, along with creating a RefID table and the ability to resolve references to the object they point to. This should make serializing and deserializing easier while laying the groundwork for optimized network syncing