Overhaul data model
As promised, the data model no longer contains direct references to other objects, instead utilizing RefIDs. Anything using RefIDs can be accessed with .Resolve(), and I plan to make that visible to the player as well. Additionally, ValueFields and ReferenceFields have been added. A ValueField contains a value and a ReferenceField contains a Reference. ValueFields inherit WorldElement, ReferenceFields do not (as they reference something which does). I will convert all instance properties in KodEngine to either Value or Reference Fields so that every slot, component, and instance property of said component and slot can be accessed with RefIDs.

More work needs to be done on the serialization end. Currently, the data model can be converted into JSON. Converting it back, however, proves to be a much tougher challenge which I plan to tackle next.