Finalized reserialization
Reserializing a world now constructs both the internal data model and the unity data model. There is a bug with character controllers where unloading and reloading the same world 3 times will lead to unexpected behavior. I plan to address this next time.

Despite that bug, this marks a huge milestone, as it is now possible to construct the state of a session over the network. While this has not been implemented, I know it will work, as it utilizes the same json deserialization that the tester method uses.

I plan to address this bug so that I can reach the next milestone: multiple people with multiple pills in a session. I believe that at this point, I can begin to demonstrate what KodVR can do.